backend/api/pyproject.toml
[tool.poetry]
name = "deplio"
version = "0.1.0"
description = ""
authors = ["Isaac Harris-Holt <isaac@withpluto.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
cadwyn = "^3.12.0"
httpx = "^0.25"
supabase = "^2.3.6"
pydantic-settings = "^2.2.1"
boto3 = "^1.34.55"
gunicorn = "^21.2.0"
uvicorn = "^0.27.1"
apscheduler = "^3.10.4"
typer = "^0.12.3"
cron-converter = "^1.1.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # __init__ files can have unused imports
